Fix build error when fetching git hash
When the parquet-rs crate is used in a third-party library, the "git"
command in the build.rs will fail. This fixes it by skipping the git
hash and just use "parquet-rs version <version-no>" as the CREATED_BY
string. This should be OK since a crate version should always map to
a unique git hash.